Made points in the all obs overlay clickable.
The strategy I adopted is far from perfect: each time an all obs tile is
requested, another request is made for JSON containing observations found
within that tile. The JSON is limited to only a few fields and is page cached,
but it's still a lot of requests and potentially a lot of data to be sending
over the wire.

An alternative strategy might be to serve and cache a custom, more compact
observation format, or to make a custom tile server that accompanies tiles
with cookies storing the obs data, a la
http://www.usnaviguide.com/ws-2008-02/demotilecookies.htm.

Todo:
  * expire the page caches