Attempt to patch a bug in with place geoms that cross the dateline.
PostGIS apparently does not handle geoms that span the dateline very well, as
in it basically considers them to be kind of inside out. Latitudes are fine,
but longitudes are flipped, so if your gem covers the Hawaiian archipelago,
finding points within it will actually select all points within the latitude
bounds that *don't* match the longitudinal bounds.

Looks like the correct approach here is to switch from GEOMETRY columns to
GEOGRAPHY columns, but that's a pretty involved task that can wait until after
Rails 3 conversion.