Fix ProjectListSpec when running multiple specs
DW.work_off defaults to 100 jobs, but that isn't always a large
enough count if there is a backlog. When ProjectListSpec was run
with enough other specs that didn't complete their work, the backlog
was growing to more than 100, causing last_observation not to get
updated in this test.

Deleting the DJ backlog before each test fixes this issue, allowing
the subspecies test to pass.