Moved faker to the test group of the Gemfile where it belongs.
This is an attempt to address an issue in which the i18n-js exports
were different across different production servers, and part of the
problem was that faker was writing translations from different locales
to that file. Not sure why it was doing so. The problem was addressed
in more recent, unreleased versions of i18n-js, but I'm hoping this
will fix it for now. If it doesn't we can revert to the checked-in
version of the translations file.