Restored statically generated translations file.
The normal translations file that i18n-js generates when assets are
compiled seems to be different depending on the server it runs on,
which leads to different fingerprints, which leads to different assets
generated on different servers, which leads to sadness (#318). Using
this static file works, but it also sucks, because it means we need to
recompile the assets and copy over the compiled translations.js every
time we update translations, which removes most of the benefit in the
i18n-js gem.