Bugfix: observation.places was broken for places crossing 180 long
The bug here is mostly related to the fact that Observation#places
tries to filter out places that don't contain the point and its
precision circle using RGeo, which seems to not do this properly for
places that cross the date line. My lame solution is to just not
perform this check for such places. Not sure if this is a bug in RGeo
or just something I couldn't figure out.