Bugfix: adding obs to project w/ fields was only adding fields
When you add to a project on obs/show and the project requires
observation fields, you see a popup prompting you to fill out the
extra fields. However, despite the fact that the fields were getting
added, the obs was not getting added to the project. I suspect this
was lost in 4328274.