Fixed #428, addressed some issues with the bulk uploader.
There was a pretty serious bug in which changing the taxon_id to that
of a threatened taxon was not preserving the original coordinates. I
think this was only happening when you did so by editing your own
observation, not when you did so by adding a new identification, or
when someone else added a new identification, so it should only have
affected a small number of observations.