Bioblitz projects can use date boundary instead of times.
Also made it so bioblitz date/time boundaries use the project owner's
time zone. The intent here is to deal with multi-day bioblitzes
spanning multiple timezones. Tying it to the owner's time zone is
problematic (things will change when the owner changes, when the owner
changes their time zone, etc), but I suspec this will solve most of
the problems.