Only set curator_coordinate_access on ProjectUser when there's opt-out.
Unless the user is presented with the option *not* to allow project
curators access to their coordinates we should assume they don't want
that. The website will still show that option as an opt-out, but in
situations like joining from the mobile apps where we may not have
added UI to control curator coordinate access, we're just assuming
they don't want that.