Tried to prevent taxon page frequency charts from updateing unecessarily.
This is really weird. React doesn't seem to recognize deep changes in properties
that are objects, and by not recognize I mean they don't even let you access
deep properties on default object props until the defaults are replaced, which
makes it kind of impossible to see if they've changed. Using shallower props
seems to address the issue, but still... dark magics.