Experimenting with previously coded taxon accumulation data.
Apparently I wrote something for this last year and never put it into a chart.
It sort of works but current has no conception of rank and treats all ancestor
taxa equally, so if you see a fox, that represents new taxa records for
everything from fox up to life. Performance didn't seem too horrible, though
that might change if/when we need to make this conscious of rank.