Replaced Observation.set_community_taxa with something that works with ES
I'm pretty sure that this method has only been accidentally working since we
separated the obs taxon from the community taxon. It was resetting the community
taxon when the user changed their opt-out preference, but since that change, it
*should* have only been changing the obs taxon. I think it actually was, but
only as a general consequence of saving the obs. It was also querying the db, so
that was making for an extremely slow delayed job.

Now it's using ES, it's user-specific, and it should only reset the obs taxon.
The observer's preferences should never affect the community taxon.