Adds a char filter to most ES analyzers for custom char mapping (#2564)
Specifically, it maps the Hawaiian ʻokina to a blank, which is how our analyzers
seems to treat apostrophes. This means searches for taxon names with the ʻokina
can be performed with the ʻokina, with an apostrophe, or without either. This also indexes the multiplication sign we use for hybrids as an x.