Only use photos in obs create if the obs doen't exist or photos were specified
Creating an obs will update an existing obs if the specified UUID belongs to an
existing obs the user posted in the past. The original intent was to accommodate
clients that lost their connection before receiving a response to a create
request and tried to create again, but we recently ran into a bug in which a
client was issueing create requests for much older observations that had photos,
and those photos were getting removed b/c these requests didn't specify any
photos and the server was assuming what the client was submitting was the
current state of the observation. This change ignores photos when creating obs
unless the obs is actually a new record or the client explicitly specified
photos to include.