Move text editor preview button to the right; make format btns untabbable
On the subject of the format buttons, this is good for those of us who use
keyboard shortcuts to tab from the taxon chooser field to the ID remarks field,
but it might be bad for people using screen readers since I think the reader
will jump over the formatting buttons. On the other hand, people using
screenreader may also want to get to the textarea sooner, so it wasn't clear to
me what was the best thing to do. FWIW, Reddit and Github also make formatting
buttons untabbable.