Reverting to react-lazy-load
reacty-lazyload supports placeholder components, which is nice, but doesn't seem
to work in the complicated scrolling context of the Compare tool, where we have
a scrollable parent element deeply nested inside an absolutely-positioned modal.
Or at least that's my theory about why it wasn't working. I tried as many of the
config props as I could and couldn't get it to work.