Removed third-party auth options on user creation form (sign up)
We want to discourage third party sign in because it's a privacy risk, it makes
it harder for us to remove or alter auth flows in the future if we can't assume
everyone has a password, it often results in people not even seeing the option
to license their content, and often leaves us with no email address we can use
to contact a user with. The buttons remain on the log in view, where new users
can still use them, but hopefully this will encourage more people to go through
with the conventional sign up process.

Also fixes a minor typo.