Attempts to make coordinate interpolation a little bit harder
When viewing an obs where you can't see the true coordinates:
  * Removed prev/next links on obs detail
  * Removed "More from" section and other extras at the bottom that show
    context
  * Truncated Date Added and Date Observed to year and month

When viewing an obs on the same day as an obs where you can't see the true
coordinates:
  * Stopped showing obscured obs as prev or next
  * Stopped showing obscured obs in contextual sections

When viewing obs/search:
  * Truncated dates when you can't see the true coordinates
  * Removed attributes that show full time stamp

On observations/:login:
  * Removed obs where you can't see the true coordinates entirely

One potential performance problem is the introduction of
coordinates_viewable_by? to observation_component. That might not be viable.