Tried to prevent using the entire places_name hash when name won't underscore
We use a key like places_name.#{place.name.parameterize.underscore} a lot, but
when the name won't underscore and we get a blank string, I18n will translate a
key like "places_name." as the entire hash of names. In addition to causing
messed up text to appear, it was really messing up the places index. This
consolidates all uses of that pattern to a single method that tries to avoid
that sitation. This is also a good example of why we should not use nested
translation keys in the future.