Crowdin should now delete existing translations when HTML files change
Normally it seems better to remove approval so proofreaders can see the old
translation and update it accordingly. This also preserves metadata like who
added the original translation. In HTML files, though, Crowdin does not track
the identity of strings by their content but by the position of their tags in
the document, which means changes to the source can associate translations with
the wrong text, leading to problems if proofreaders approve translations en
masse. Deleting the translations isn't great, but it will prevent nonsensical
translations from getting through.