Bugfix: announcements form wasn't working when locale was Luxembourgish
This is kind of a work around for what looks like a Rails bug, where they are
looking for a translatable date.order string, but raising an exception when it
contains anything other than the English values year, month, and day. I guess
the order is translatable but the values aren't which seems very confusing.