Run all JS dependencies through Babel during build
As of c8b403b4e9bd5fffc359aaa7cc086162fc068dd7, some of our dependencies are in
ES6 or higher. Previously, we had configured webpack to ignore everything in
node_modules during it's Babel step because everything we were using was already
transformed into more browser-compliant forms of JS, but since we can no longer
assume that is true, we need to run everything through Babel so if a dependency
uses ES6, we perform our own transformations to browser-compliant JS. This slows
things down and seems like it might increase the size of the JS assets. If we
don't do this, we end up running a JS file with ES6 syntax through Uglifier
during asset precompilation, which gives it indigestion. Uglifier *can* be
configured to handle ES6 with `harmony: true`, but trying that led to other
problems with the version of Uglifier we're using that may not have been
resolved (see https://github.com/lautis/uglifier/issues/137).

Note that I tried experimenting with the `exclude` feature in the webpack config
so it wouldn't exclude certain directories in node_modules, but it's really hard
to figure out exactly what dependency is using some piece of ES6 syntax that
causes.