Update it.yml
Removing redundant indefinite article that produced "una una osservazione" in the rendered HTML when substituting for %{noun}. 

Note that the use of a feminine indefinite article "una" preceding %{x} in the strings starting at line 4522 on creates an issue when the value is masculine (e.g. "una commento") and there's no easy way to fix this given how these messages are structured. It looks like the previous translator attempted to address this with (what I believe should read) "un/una' on line 4524-4525.