Custom tokens controller was broken in Rails 5
We had some custom stuff to prevent suspended users from retrieving an access
token that stopped working in updated versions of Doorkeeper (or Devise). This
should fix that, plus supports the same block when a user attempts to obtain a
token via any authentication flow, not just the Resource Owner Password
Credentials flow.