Fix paragraphs not being applied in journal posts
Fixes inaturalist/#3204

Paragraphs were not being applied when in between 2 sections of tables,
lists or preformatted text.
This was because the regex used was greedy, and would  include
everything between the first and last of one of those tags, rather than
stop matching after the first end tag it finds.

Fix was to apply the non-greedy `?` modifier to the `*`s