Disable CSRF protection when authenticated w/ OAuth or JWT
Fixes a bug in which OAuth- and JWT-authenticated requests were signing people
out of their cookie-based sessions, or rather expiring those sessions so they
would not be "remembered" on subsequent visits to the website. Something as
simple as uploading an observation in the Uploader could trigger this.

Disabling forgery protection in these cases might represent a security risk, but
an attacker who has an existing access token can already use that token in much
simpler attacks, so it's not clear to me how this would make that situation
worse.