More attempts to deal with CSRF problems (#3280)
* Attempts to make specs a little more idempotent

* Skip CSRF protection if an Authorization header is present

CSRF protection was returning 500s when users were submitting invalid or expired
JWTs, when they should have been getting 401s. This skips CSRF protection if
the Authorization header is specified at all: if the specified token is bad,
that should get handled in the appropraite auth callbacks, not by CSRF
protection.

In theory this should not pose a security risk because the Authorization header
can't be set in an HTML form, and even if it can, auth should fail unless the
attacker has a valid access token, which would enable them to perform simpler
and more effective attacks than CSRF.

* Disabled action caching for taxa/show; fixed some boolean nonsense

* Removed access_token param authentication

We don't document support for it and we don't use it, so I'm not sure why we
support it.