Fix taxon photo term filter (#3291)
So both term_value_id AND term_id need to be valid for the taxon for
both of them to be included. Fixes #3250
Beforehand, if the term_id was valid, but not the term_value_id, then
the term_id was included in the filter which would filter out all
observations that didn't include some value for that term_id, but
the UI would still be indicating that no filtering was occurring.