Changed namespaces and removed establishmentMeans from DwC-A
As of https://lists.gbif.org/pipermail/api-users/2022-January/000643.html,
these are the namespaces we should be using for recordedById and
identifiedById. establishmentMeans now has an official controlled vocab that
makes it clear we've been populating that field incorrectly all along. The
new conventional way to report something similar to our captive/cultivated
status is to use the `pathway` and `degreeOfEstablishment` fields, but
unfortunately we don't record information with enough granularity to fill
those out at present.