Registration via API should return a JSON for an unconfirmed user
But not JSON that describes that user indicating successful sign in. This just
returns JSON that includes a message about needing to confirm the email
address, which may be a signicicant change for some clients that assume
successful registration means the user can authenticate and, in fact, is
authenticated, which is no longer the case with the need for email
confirmation.