Bugfix: uploader was not correctly interpreting time zone from the server (#3433)
Most of the time the uploader will correctly interpret the datetime from the
EXIF and populate that value in the obs card immediately. However, if it
fails, it will fall back on the `time_observed_at` provided by the server.
The bug was occurring when the latter occurred and the uploader was correctly
interpreting that value with its specified offset, but then incorrectly
formatting that in the *local* time zone for display. This change makes sure
that when the client interprets `time_observed_at`, it also sets the moment
object to use the correct time zone before formatting the datetime for
display.