RTL support for /identifications/:login, signup, and signin (#3577)
Starts support for right-to-left layouts (RTL) for languages like Arabic, Farsi, and Hebrew by overriding styles in extra stylesheets included for users viewing the site in RTL languages. As of this PR this is a test that can be enabled on signup, signin, and /identifications/:login with the `test=rtl` param.