Remove facebook (#3748)
* Remove all Facebook functionality (#3720)
* Removed FacebookPhoto and Facebook assertion flow

There are no remaining FacebookPhoto records in production, so that model is probably safe to remove. We don't store Facebook tokens, so there's no reason to support the assertion flow since we can neither get a new token from Facebook or compare a token with one we have on file.