Move interpolate link to map
I figure it will be more obvious there, plus it's a bit easier to determine
whether or not to show it.

Also added better support for the not_id param in Observation.elastic_query,
and altered Observation#interpolate_coordinates so it can handle situations
where the prev and next obs have identifical times.