refactor: split DQA help into smaller strings (#4309)
Previously we've kept this string large to provide the maximum amount of
context for translators, and changed the key every time we updated the text,
requiring them to re-translate everything on the assumption that Crowdin's
Translation Memory feature will show the previous translation. We've recently
learned there is an unknown limit on the size of string TM will work for,
which changes the context/change resilience tradeoff in favor of smaller
strings.