fix: redeliver unsuspended user's messages whenever they were suspended
The previous strategy avoided redelivering messages that were intentionally
deleted by the recipient with a kind of statute of limitations approach. This
approach tracks whether a message was ever sent with an additional column in
the database that we unset when we suspend the sender and delete the
recipient copies, but we don't unset when the recipient manually deletes
their copy.

This introduces some confusion b/c we cannot set sent_at for messages created
before this gets released, because there is no way to know if a message has
no to_user_copy because of suspension or because of manual deletion.