fix: localize GMaps using browser locale if no region in user locale (#4392)
Google assumes the region is US unless we specify it, but many of our locales
don't specify a region. This tries to get the region from the locale
specified in the HTTP request (usually from the browser) so we can increase
the chance that users will get appropriately localized map tiles... or at
least as appropriate as Google sees fit.