fix: skip consent for subsequent google signins (#4389)
Our previous omniauth config was forcing the user to grant consent to
permissions *every* time they signed in. In theory, this should only prompt
for that consent the first time, and subsequent attempts to sign in should
present an option to select a Google account and then take them directly to
iNat.