fix: restore standard behavior of defaultValue in I18n.js
We had overridden this option so that the default value would be used even if
fallback translations existed, which is *not* how it works in Rails, and is
frankly just kind of confusing. Instead, I've changed the option to
defaultValuePreFallback to be explicit about the fact that it gets used
before even checking for fallback translations.

FWIW, this is meant for use in situations where an alternate translation
exists in the current locale that would be better than the default English
fallback that exists for all locales.