fix: stop dashboard updates after user unsubscribed from resource
The dashboard stubs out UpdateActions for additional content associated with
resources that might generate notifications, e.g. if you are subscribed to a
journal post and an UpdateAction exists for a comment on that post, you will
both be notified about it on your dashboard *and* all the additional comments
on that post will be loaded as well, so you can see the full context.
However, we were doing that loading even if you had unsubscribed from the
post, so in situations where you were subscribed, a comment was made and thus
an UpdateAction existed for you, but then you unsubscribed from the post, we
were showing you that UpdateAction as well as all the subsequent comments,
which made it seem like you were still subscribed to subsequent updates. This
stops that additional loading if you've unsubscribed.